rename use-modal
